# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cash-Track API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://cash-track.app
  description: |
    Cash-Track is a personal finance management SaaS. This API powers the mobile and web clients.

    ## Authentication
    Most endpoints require a **JWT Bearer token** obtained via `/v1/auth/login`, `/v1/auth/register`,
    or one of the social/passkey login flows.

    Tokens come in pairs:
    - **Access token** — short-lived (~1 hour). Send in `Authorization: Bearer <token>`.
    - **Refresh token** — long-lived (~7 days). Use `POST /v1/auth/refresh` to get a new pair.

    ## Email Confirmation
    After registration the user receives a confirmation email. Creating wallets, charges, and tags
    requires a confirmed email address. Use the `/v1/auth/email/confirmation` group to check status
    or resend the link.

    ## Error Responses
    Standard errors return an `Error` object. Validation failures return a `ValidationError`
    object with per-field messages (HTTP 422).

    ## Idempotency
    Most state-changing endpoints (`POST`/`PUT`/`PATCH`/`DELETE`) accept an optional
    `Idempotency-Key` request header — a client-generated UUIDv4. Retrying the exact same
    request with the same key within 24 hours replays the original response instead of
    re-executing it (`Idempotency-Replayed: true` is added to the replayed response). Reusing
...
    rejected (`400`). See `IdempotencyKeyParam`, `IdempotencyKeyConflict`, and
...

...

...

...
servers:
...
paths:
...
  /v1/wallets/{id}/limits:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"

    get:
      operationId: limitList
      summary: List wallet limits
      description: |
        Returns all budget limits for the wallet. Each limit includes the current
        expenditure as a percentage of the limit amount.
      tags: [Limits]
      security:
        - bearerAuth: []
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Wallet limits with progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletLimit"

    post:
      operationId: limitCreate
      summary: Create a budget limit
      description: |
        Creates a new budget limit for a wallet from one or more tag groups.
        Requires at least one tag group. The limit fires when the summed
        charges matching each group's tags exceed `amount`.
      tags: [Limits]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLimitRequest"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Created limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Limit"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
  /v1/wallets/{walletId}/limits/{limitId}:
    parameters:
      - $ref: "#/components/parameters/WalletIdQueryParam"
      - $ref: "#/components/parameters/LimitIdParam"

    put:
      operationId: limitUpdate
      summary: Update a limit
      description: Replaces all fields of an existing budget limit.
      tags: [Limits]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLimitRequest"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Updated limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Limit"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
    delete:
      operationId: limitDelete
      summary: Delete a limit
      description: Permanently removes a budget limit.
      tags: [Limits]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Limit deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
        "422":
          $ref: "#/components/responses/IdempotencyKeyMismatch"
  /v1/wallets/{walletId}/limits/copy/{sourceWalletId}:
    post:
      operationId: limitCopy
      summary: Copy limits from another wallet
      description: |
        Copies all budget limits from `sourceWalletId` into the target wallet.
        Existing limits in the target wallet are not removed. Tags are matched by name.
      tags: [Limits]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WalletIdQueryParam"
        - name: sourceWalletId
          in: path
          required: true
          description: Wallet to copy limits from
          schema:
            type: integer
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Limits copied
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
        "422":
          $ref: "#/components/responses/IdempotencyKeyMismatch"
  # =========================================================================
  # TAGS
  # =========================================================================
  /v1/tags:
    get:
      operationId: tagList
      summary: List user tags
      description: Returns all tags owned by the authenticated user.
      tags: [Tags]
      security:
        - bearerAuth: []
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: User's tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"

    post:
      operationId: tagCreate
      summary: Create a tag
      description: |
        Creates a new tag for the current user. Requires confirmed email.
        Tag names must be unique per user, 3–255 characters, and contain no spaces.
      tags: [Tags]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
            example:
              name: food
              icon: "🍔"
              color: "#FF5733"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/v1/auth/login`, `/v1/auth/register`, or any social
        login endpoint. Tokens expire after ~1 hour. Refresh via `POST /v1/auth/refresh`.

  # -------------------------------------------------------------------------
  # PARAMETERS
  # -------------------------------------------------------------------------
  parameters:
    WalletIdParam:
      name: id
      in: path
      required: true
      description: Wallet integer ID
      schema:
        type: integer
        minimum: 1
    WalletIdQueryParam:
      name: walletId
      in: path
      required: true
      description: Wallet integer ID
      schema:
        type: integer
        minimum: 1
...
    LimitIdParam:
      name: limitId
      in: path
      required: true
      description: Limit integer ID
      schema:
        type: integer
        minimum: 1
...
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional. A canonical lowercase UUIDv4 supplied by the client. Sending the same key
        again with the same method, path, and request body within 24 hours replays the first
        response verbatim (see the `Idempotency-Replayed` response header) instead of
        re-executing the request. Sending it again with a *different* body returns `422`
        instead of running — see the `IdempotencyKeyMismatch` response. A malformed value
        (not a canonical UUIDv4) returns `400`. Omitting the header entirely runs the request
        normally, with no deduplication.
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
        example: "1e6c2a6a-9e2e-4d3a-8f2e-6a2f1b8c9d10"

  # -------------------------------------------------------------------------
  # SCHEMAS
  # -------------------------------------------------------------------------
  schemas:
...
    WalletShort:
      type: object
      description: Minimal wallet representation used as a nested field inside charges and limits.
      required:
        - type
        - id
        - name
        - slug
        - totalAmount
        - isActive
        - isPublic
        - isArchived
        - createdAt
        - updatedAt
      properties:
        type:
          type: string
          enum: [walletShort]
          readOnly: true
        id:
          type: integer
          readOnly: true
        name:
          type: string
        slug:
          type: string
        totalAmount:
          type: number
          format: float
        isActive:
          type: boolean
        isPublic:
          type: boolean
        isArchived:
          type: boolean
        defaultCurrencyCode:
          type: [string, "null"]
        defaultCurrency:
          $ref: "#/components/schemas/Currency"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Tag:
      type: object
      description: A label attached to charges for categorisation.
      required:
        - type
        - id
        - name
        - userId
        - createdAt
        - updatedAt
      properties:
        type:
          type: string
          enum: [tag]
          readOnly: true
        id:
          type: integer
          readOnly: true
        name:
          type: string
          minLength: 3
          maxLength: 255
          description: No spaces allowed; unique per user
        icon:
          type: [string, "null"]
          description: Emoji or short string (1–10 characters)
          example: "🍔"
        color:
          type: [string, "null"]
          pattern: "^#[0-9A-Fa-f]{6}$"
          description: Hex color code
          example: "#FF5733"
        userId:
          type: integer
          description: Owner user ID
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    CreateTagRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
          pattern: "^\\S+$"
          description: Must be unique per user and contain no spaces
        icon:
          type: [string, "null"]
          minLength: 1
          maxLength: 10
          description: Emoji or short symbol
        color:
          type: [string, "null"]
          pattern: "^#[0-9A-Fa-f]{6}$"
          description: Hex color code (e.g. `#FF5733`)

    # -----------------------------------------------------------------------
    # LIMIT
    # -----------------------------------------------------------------------
    LimitTagGroup:
      type: object
      description: |
        A group of tags within a limit. Charges are matched against each group
        independently using its `connection` (`and` requires every tag in the
        group to be present on a charge, `or` requires any one of them); the
        limit's totals are the sum across all its groups, without deduplicating
        charges matched by more than one group.
      required:
        - type
        - connection
        - tags
      properties:
        type:
          type: string
          enum: [limitTagGroup]
          readOnly: true
        connection:
          type: string
          enum: ["and", "or"]
          description: How this group's tags are matched against a charge's tags
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
    Limit:
      type: object
      description: A budget ceiling for charges matching a set of tag groups in a wallet.
      required:
        - type
        - id
        - operation
        - amount
        - walletId
        - createdAt
        - updatedAt
        - tags
        - tagGroups
      properties:
        type:
          type: string
          enum: [limit]
          readOnly: true
        id:
          type: integer
          readOnly: true
        operation:
          type: string
          enum: ["+", "-"]
          description: "`+` = income limit, `-` = expense limit"
        amount:
          type: number
          format: float
          minimum: 0.01
        walletId:
          type: integer
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
          deprecated: true
          description: Deprecated, always empty. Use `tagGroups` instead.
        tagGroups:
          type: array
          items:
            $ref: "#/components/schemas/LimitTagGroup"
        wallet:
          $ref: "#/components/schemas/WalletShort"
    WalletLimit:
      type: object
      description: A limit with its current spend as a percentage.
      required:
        - amount
        - percentage
        - limit
      properties:
        amount:
          type: number
          format: float
          description: Current amount spent / earned matching this limit's tags
        percentage:
          type: number
          format: float
          description: "Current spend as a fraction of the limit (0–1+). Values > 1 mean the limit is exceeded."
          example: 0.75
        limit:
          $ref: "#/components/schemas/Limit"
    CreateLimitTagGroupRequest:
      type: object
      required:
        - operation
        - tags
      properties:
        operation:
          type: string
          enum: ["and", "or"]
          description: How this group's tags are matched against a charge's tags
        tags:
          type: array
          minItems: 1
          items:
            type: integer
          description: Tag IDs in this group (min 1)
    CreateLimitRequest:
      type: object
      required:
        - type
        - amount
        - tagGroups
      properties:
        type:
          type: string
          enum: ["+", "-"]
          description: "`+` = income limit, `-` = expense limit"
        amount:
          type: number
          format: float
          minimum: 0.01
        tags:
          type: array
          items:
            type: integer
          deprecated: true
          description: Deprecated and ignored. Use `tagGroups` instead.
        tagGroups:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateLimitTagGroupRequest"
          description: |
            Tag groups this limit applies to (min 1). Each group's tags are
            matched using its own `operation`; totals are summed across groups
            without deduplicating charges matched by more than one group.

    # -----------------------------------------------------------------------
    # CURRENCY
    # -----------------------------------------------------------------------
    Currency:
      type: object
      description: A supported currency with live exchange rate.
      required:
        - type
        - id
        - code
        - name
        - char
        - rate
        - updatedAt
      properties:
        type:
          type: string
          enum: [currency]
          readOnly: true
        id:
          type: string
          description: ISO 4217 code (same as `code`)
          example: USD
        code:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code
          example: USD
        name:
          type: string
          description: Full currency name
          example: US Dollar
        char:
          type: string
          description: Currency symbol
          example: "$"
        rate:
          type: number
          format: float
          description: Exchange rate relative to the base currency (USD)
        updatedAt:
          type: string
          format: date-time
          description: When the exchange rate was last updated

    # -----------------------------------------------------------------------
    # PAGINATION
    # -----------------------------------------------------------------------
...
    Error:
      type: object
      description: Standard error envelope.
      required:
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code (optional)
        message:
          type: string
          description: Human-readable error description
    ValidationError:
      type: object
      description: Field-level validation failures (HTTP 422).
      required:
        - errors
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of field names to arrays of error messages
          example:
            email:
              - "Must be a valid email address"
            password:
              - "Must be at least 6 characters"

  # -------------------------------------------------------------------------
  # REUSABLE HEADERS
  # -------------------------------------------------------------------------
  headers:
    ApiVersion:
      description: |
        Git tag of the running API image (e.g. `v1.4.2`). Present on every response; omitted
        entirely (rather than sent empty) when the image was built without a tag.
      schema:
        type: string
    ApiSha:
      description: |
        Git commit SHA of the running API image. Present on every response; omitted entirely
        (rather than sent empty) when the image was built without a commit SHA.
      schema:
        type: string
    TraceId:
      description: |
        The active OpenTelemetry trace ID. Present on every response; omitted entirely (rather
        than sent empty or all-zeros) when tracing is disabled (no valid active span context).
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
    IdempotencyReplayed:
      description: |
        `true` when this response is a verbatim replay of the first request that used this
        `Idempotency-Key`, rather than a fresh execution. Absent on a fresh response.
      schema:
        type: string
        enum: ["true"]

  # -------------------------------------------------------------------------
  # REUSABLE RESPONSES
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: |
        Bad request — invalid input or business rule violation. Also returned when an
        `Idempotency-Key` header is present but not a canonical lowercase UUIDv4.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: INVALID_CREDENTIALS
            message: The provided credentials are incorrect.
    Unauthorized:
      description: Authentication required or token expired
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Unauthenticated.
    NotFound:
      description: Resource not found
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Resource not found.
    IdempotencyKeyConflict:
      description: |
        A request with this `Idempotency-Key` (same scope, method, path, and request body) is
        still being processed. Retry after the given delay; do not send a different body under
        the same key while this is in flight.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: The original request with this Idempotency-Key is still processing. Please retry shortly.
    IdempotencyKeyMismatch:
      description: |
        This `Idempotency-Key` was already used with a different request body (method, path, or
        payload). Idempotency keys are single-use per distinct request — retry with a new key,
        or resend the exact original request body to get the stored result.

        Deliberately `Error`-shaped (a plain `message`, no `errors` map) rather than
        `ValidationError`-shaped: this is a conflict over key reuse, not a field-level
        validation failure, and callers should not try to render it as one.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: This Idempotency-Key was already used with a different request body.
...
    ValidationOrIdempotencyMismatch:
      description: |
        `422` on this endpoint has two distinct causes, both possible since it accepts an
        `Idempotency-Key`:
          - **Field-level validation failure** — `ValidationError`-shaped, carries an `errors`
            map keyed by field name.
          - **Idempotency key reused with a different body** — `Error`-shaped, a plain
            `message`, no `errors` key.

        Discriminate by the presence of `errors` in the body, not by anything else — both
        causes share this single `422` status code.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ValidationError"
              - $ref: "#/components/schemas/Error"
...
security:
  - bearerAuth: []